Load the expression printer on demand
