+++ frontend/src/metabase/querying/expressions/formatter/index.ts
@@ -1,1 +1,2 @@
 export * from "./formatter";
+export * from "./types";
