+++ frontend/src/metabase/querying/expressions/formatter/print.ts
@@ -1,0 +1,358 @@
+import { builders, printer } from "prettier/doc";
+
+import { parseNumber } from "metabase/utils/number";
+import * as Lib from "metabase-lib";
+import { isa } from "metabase-lib/v1/types/utils/isa";
+
+import { getClauseDefinition } from "../clause";
+import {
+  formatDimensionName,
+  formatIdentifier,
+  formatMetricName,
+  formatSegmentName,
+} from "../identifier";
+import { parsePunctuator } from "../punctuator";
+import { formatStringLiteral } from "../string";
+
+import type { ExpressionNode, FormatOptions } from "./types";
+import {
+  Path,
+  pathMatchers as check,
+  isDimensionOperator,
+  isExpressionOperator,
+  isValueOperator,
+} from "./utils";
+
+type Doc = builders.Doc;
+
+const {
+  // prettier helpers
+  join,
+  indent,
+  softline,
+  line,
+  group,
+  ifBreak,
+} = builders;
+
+const DEFAULT_PRINT_WIDTH = 80;
+const TAB_WIDTH = 2;
+
+export function printExpressionParts(
+  root: ExpressionNode,
+  options: FormatOptions,
+): string {
+  const doc = print(new Path(root), options);
+  const { formatted } = printer.printDocToString(doc, {
+    printWidth: options.printWidth ?? DEFAULT_PRINT_WIDTH,
+    tabWidth: TAB_WIDTH,
+  });
+  return formatted;
+}
+
+function print(path: Path<ExpressionNode>, options: FormatOptions): Doc {
+  if (path.node === null) {
+    return "";
+  } else if (check.isNumberLiteral(path)) {
+    return formatNumberLiteral(path.node);
+  } else if (check.isBooleanLiteral(path)) {
+    return formatBooleanLiteral(path.node);
+  } else if (check.isStringLiteral(path)) {
+    return formatStringLiteral(path.node, options.stringDelimiter);
+  } else if (check.isColumnMetadata(path)) {
+    return formatColumn(path, options);
+  } else if (check.isMetricMetadata(path)) {
+    return formatMetric(path, options);
+  } else if (check.isMeasureMetadata(path)) {
+    return formatMeasure(path, options);
+  } else if (check.isSegmentMetadata(path)) {
+    return formatSegment(path, options);
+  } else if (check.isExpressionOperator(path)) {
+    return formatOperator(path, options);
+  } else if (check.isDimensionOperator(path)) {
+    return formatDimension(path);
+  } else if (check.isValueOperator(path)) {
+    return formatValueExpression(path, options);
+  } else if (check.isExpressionParts(path)) {
+    return formatFunctionCall(path, options);
+  }
+
+  throw new Error(`Unknown MBQL clause: ${JSON.stringify(path.node)}`);
+}
+
+// Helper to recurse into an AST node that is not easily expressed as a
+// property of the node currently in path.
+function recurse(
+  path: Path<ExpressionNode>,
+  options: FormatOptions,
+  node: ExpressionNode,
+): Doc {
+  const { length } = path.stack;
+  path.stack.push(node);
+  try {
+    return print(path, options);
+  } finally {
+    path.stack.length = length;
+  }
+}
+
+function formatNumberLiteral(node: number | bigint): Doc {
+  return String(node);
+}
+
+function formatBooleanLiteral(node: boolean): Doc {
+  return node ? "True" : "False";
+}
+
+function assert(condition: any, msg: string): asserts condition {
+  if (!condition) {
+    throw new Error(msg);
+  }
+}
+
+function formatColumn(
+  path: Path<Lib.ColumnMetadata>,
+  options: FormatOptions,
+): Doc {
+  const { query, stageIndex } = options;
+  const column = path.node;
+
+  assert(query !== undefined, "Expected query");
+  assert(typeof stageIndex === "number", "Expected stageIndex");
+  assert(Lib.isColumnMetadata(column), "Expected column");
+
+  const info = Lib.displayInfo(query, stageIndex, column);
+  return formatDimensionName(info.longDisplayName);
+}
+
+function formatMetric(
+  path: Path<Lib.MetricMetadata>,
+  options: FormatOptions,
+): Doc {
+  const metric = path.node;
+  const { query, stageIndex } = options;
+
+  assert(query !== undefined, "Expected query");
+  assert(typeof stageIndex === "number", "Expected stageIndex");
+  assert(Lib.isMetricMetadata(metric), "Expected metric");
+
+  const displayInfo = Lib.displayInfo(query, stageIndex, metric);
+  return formatMetricName(displayInfo.displayName);
+}
+
+function formatMeasure(
+  path: Path<Lib.MeasureMetadata>,
+  options: FormatOptions,
+): Doc {
+  const metric = path.node;
+  const { query, stageIndex } = options;
+
+  assert(query !== undefined, "Expected query");
+  assert(typeof stageIndex === "number", "Expected stageIndex");
+  assert(Lib.isMeasureMetadata(metric), "Expected measure");
+
+  const displayInfo = Lib.displayInfo(query, stageIndex, metric);
+  return formatMetricName(displayInfo.displayName);
+}
+
+function formatSegment(
+  path: Path<Lib.SegmentMetadata>,
+  options: FormatOptions,
+) {
+  const segment = path.node;
+  const { query, stageIndex } = options;
+
+  assert(query !== undefined, "Expected query");
+  assert(typeof stageIndex === "number", "Expected stageIndex");
+  assert(Lib.isSegmentMetadata(segment), "Expected segment");
+
+  const displayInfo = Lib.displayInfo(query, stageIndex, segment);
+  return formatSegmentName(displayInfo.displayName);
+}
+
+function formatDimension(path: Path<Lib.ExpressionParts>): Doc {
+  const { node } = path;
+  assert(isDimensionOperator(node), "Expected dimension");
+
+  const name = node.args[0];
+  assert(typeof name === "string", "Expected expression name to be a string");
+
+  return formatIdentifier(name);
+}
+
+function formatFunctionCall(
+  path: Path<Lib.ExpressionParts>,
+  options: FormatOptions,
+): Doc {
+  const { node } = path;
+
+  const args = node.args.map((arg: ExpressionNode) =>
+    recurse(path, options, arg),
+  );
+
+  const expressionOptions = formatExpressionOptions(node.options);
+  if (expressionOptions) {
+    args.push(expressionOptions);
+  }
+
+  // render a call expression as
+  //
+  //   callee(arg1, arg2, ...)
+  //
+  // or, when not enough space
+  //
+  //   callee(
+  //     arg1,
+  //     arg2,
+  //     ...
+  //   )
+  return group([
+    displayName(node.operator),
+    "(",
+    indent([
+      // indent args
+      softline,
+      join([",", line], args),
+    ]),
+    softline,
+    ")",
+  ]);
+}
+
+function formatExpressionOptions(options: Lib.ExpressionOptions): Doc | null {
+  if ("case-sensitive" in options && !options["case-sensitive"]) {
+    return formatStringLiteral("case-insensitive");
+  }
+  if ("include-current" in options && options["include-current"]) {
+    return formatStringLiteral("include-current");
+  }
+  if ("mode" in options && options.mode) {
+    const camelCased = options.mode.replace(/-\w/g, (match) =>
+      match.charAt(1).toUpperCase(),
+    );
+    return formatStringLiteral(camelCased);
+  }
+  return null;
+}
+
+function formatOperator(
+  path: Path<Lib.ExpressionParts>,
+  options: FormatOptions,
+): Doc {
+  const { node } = path;
+
+  assert(
+    isExpressionOperator(node),
+    `Expected operator but got ${node.operator}`,
+  );
+
+  const operatorGoesOnNewline =
+    node.operator === "and" || node.operator === "or";
+
+  const args = node.args.map((arg, index) => {
+    const ln = index === 0 || operatorGoesOnNewline ? "" : line;
+
+    function ind(doc: Doc) {
+      if (index === 0 || operatorGoesOnNewline) {
+        return doc;
+      }
+      return indent(doc);
+    }
+
+    if (
+      !Lib.isExpressionParts(arg) ||
+      isValueOperator(arg) ||
+      isDimensionOperator(arg)
+    ) {
+      // Not a call expression so not an operator
+      return ind([ln, recurse(path, options, path.node.args[index])]);
+    }
+
+    const argOperator = arg.operator;
+    const formattedArg = recurse(path, options, path.node.args[index]);
+
+    const operatorPrecedence = precedence(node.operator);
+    const argPrecedence = precedence(argOperator);
+
+    const isLowerPrecedence = operatorPrecedence > argPrecedence;
+    const isSamePrecedence = operatorPrecedence === argPrecedence;
+
+    // "*","/" always have two arguments.
+    // If the second argument of "/" is an expression, we have to calculate it first.
+    // Hence, adding parenthesis.
+    // "a / b * c" vs "a / (b * c)", "a / b / c" vs "a / (b / c)"
+    // "a - b + c" vs "a - (b + c)", "a - b - c" vs "a - (b - c)"
+    const isSamePrecedenceWithExecutionPriority = index > 0 && isSamePrecedence;
+
+    const shouldUseParens =
+      isLowerPrecedence || isSamePrecedenceWithExecutionPriority;
+
+    if (shouldUseParens) {
+      return ind(
+        group([
+          ln,
+          "(",
+          group(
+            ifBreak(
+              [indent([softline, group(formattedArg)]), softline],
+              formattedArg,
+            ),
+          ),
+          ")",
+        ]),
+      );
+    } else {
+      return ind([ln, formattedArg]);
+    }
+  });
+
+  if (isUnaryOperator(node.operator)) {
+    return [displayName(node.operator), " ", args[0]];
+  }
+
+  if (operatorGoesOnNewline) {
+    return group(join([line, displayName(node.operator), " "], args));
+  } else {
+    return group(join([" ", displayName(node.operator)], args));
+  }
+}
+
+function precedence(op: string): number {
+  return parsePunctuator(op)?.precedence ?? Infinity;
+}
+
+function isUnaryOperator(op: Lib.ExpressionOperator) {
+  return getClauseDefinition(op)?.args.length === 1;
+}
+
+function formatValueExpression(
+  path: Path<Lib.ExpressionParts>,
+  options: FormatOptions,
+): Doc {
+  const { node } = path;
+
+  assert(isValueOperator(node), "Expected value");
+
+  const {
+    args: [value],
+    options: valueOptions,
+  } = node;
+
+  const baseType = valueOptions?.["base-type"];
+  if (
+    typeof value === "string" &&
+    typeof baseType === "string" &&
+    isa(baseType, "type/BigInteger")
+  ) {
+    const number = parseNumber(value);
+    if (number != null) {
+      return recurse(path, options, number);
+    }
+  }
+  return recurse(path, options, node.args[0]);
+}
+
+function displayName(name: Lib.ExpressionOperator): string {
+  assert(name !== "value", "Unexpected value clause");
+  return getClauseDefinition(name)?.displayName ?? name;
+}
